+++ src/main/java/com/coro/coro/member/service/MemberService.java
@@ -1,12 +1,6 @@
 package com.coro.coro.member.service;
 
-import com.coro.coro.auth.domain.RefreshToken;
-import com.coro.coro.auth.dto.response.TokenResponse;
-import com.coro.coro.auth.jwt.JwtProvider;
-import com.coro.coro.auth.service.port.RefreshTokenRepository;
-import com.coro.coro.common.service.port.UUIDHolder;
 import com.coro.coro.member.domain.Member;
-import com.coro.coro.member.dto.request.MemberLoginRequest;
 import com.coro.coro.member.dto.request.MemberModificationRequest;
 import com.coro.coro.member.dto.request.MemberRegisterRequest;
 import com.coro.coro.member.exception.MemberException;
@@ -30,11 +24,8 @@ import static com.coro.coro.common.response.error.ErrorType.*;
 @Builder
 public class MemberService {
 
-    private final JwtProvider tokenProvider;
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final UUIDHolder uuidHolder;
-    private final RefreshTokenRepository refreshTokenRepository;
 
     /* 회원가입 */
     @Transactional
@@ -52,29 +43,6 @@ public class MemberService {
         member.encryptPassword(passwordEncoder);
 
         return memberRepository.save(member);
-    }
-
-    /* 로그인 */
-    public TokenResponse login(final MemberLoginRequest requestMember) {
-        Member member = getMemberByEmail(requestMember);
-        boolean isRightPassword = member.isRightPassword(requestMember.getPassword(), passwordEncoder);
-        if (!isRightPassword) {
-            throw new MemberException(MEMBER_NOT_VALID_PASSWORD);
-        }
-
-        String accessToken = tokenProvider.generateAccessToken(member.getNickname());
-        String refreshToken = uuidHolder.generateUUID();
-        RefreshToken token = RefreshToken.builder()
-                .nickname(member.getNickname())
-                .refreshToken(refreshToken)
-                .build();
-        refreshTokenRepository.save(token);
-        return new TokenResponse(accessToken, refreshToken);
-    }
-
-    private Member getMemberByEmail(final MemberLoginRequest requestMember) {
-        return memberRepository.findByEmail(requestMember.getEmail())
-                .orElseThrow(() -> new MemberException(MEMBER_NOT_VALID_EMAIL));
     }
 
     /* 회원 수정 */
